% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAFE-package.R
\docType{package}
\name{CCAFE-package}
\alias{CCAFE-package}
\title{CCAFE: Case Control Allele Frequency Estimation}
\description{
Functions to reconstruct case and control AFs from summary statistics. One function uses OR, NCase, NControl, and SE(log(OR)). The second function uses OR, NCase, NControl, and AF for the whole sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wolffha/CCAFE/}
  \item Report bugs at \url{https://github.com/wolffha/CCAFE/issues}
}

}
\author{
\strong{Maintainer}: Hayley Wolff \email{hayley.wolff@cuanschutz.edu}

}
\keyword{internal}
