% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAFE-package.R
\name{CCAFE}
\alias{CCAFE}
\title{CCAFE: Case and Control Allele Frequency Estimation}
\description{
CCAFE: Case and Control Allele Frequency Estimation
}
\section{Main functions}{

\itemize{
\item \code{\link[=CaseControl_AF]{CaseControl_AF()}}
\item \code{\link[=CaseControl_SE]{CaseControl_SE()}}
}
}

\section{Utility functions}{

\itemize{
\item \code{\link[=CCAFE_convertVCF]{CCAFE_convertVCF()}}
}
}

